[feat/redesign] v2 version, move files to under shared and versions/v1 for old codes
