+++ src/_theme.scss
@@ -2,7 +2,7 @@
 
 $colors: (
   singapore-malaysia-2024: (
-    base: #2b2e4a, // need to be in sync with utils/const.ts
+    base: #2b2e4a, // need to be in sync with shared/utils/const.ts
     base2: #f8c471,
     button: #27ae60,
     button-hover: #1f7a4d,
@@ -10,31 +10,31 @@ $colors: (
     caption-color: #a3b0c6,
   ),
   refreshing-bali-2020: (
-    base: #d1e4e2, // need to be in sync with utils/const.ts
+    base: #d1e4e2, // need to be in sync with shared/utils/const.ts
     base2: #f4e1c1,
     button: #d2691e,
     button-hover: #8b4513
   ),
   luxury-bali-2019: (
-    base: #f5f5dc, // need to be in sync with utils/const.ts
+    base: #f5f5dc, // need to be in sync with shared/utils/const.ts
     base2: #87ceeb,
     button: #6b8e23,
     button-hover: #556b2f
   ),
   spring-japan-2018: (
-    base: #ffd7e8, // need to be in sync with utils/const.ts
+    base: #ffd7e8, // need to be in sync with shared/utils/const.ts
     base2: #a2c8f7,
     button: #c2185b,
     button-hover: #a3154a
   ),
   ora-beach-2018: (
-    base: #a0d8ef, // need to be in sync with utils/const.ts
+    base: #a0d8ef, // need to be in sync with shared/utils/const.ts
     base2: #ffd1b3,
     button: #4a90e2,
     button-hover: #3b7bbf,
   ),
   explore-singapore-2018: (
-    base: #f0f0f0, // need to be in sync with utils/const.ts
+    base: #f0f0f0, // need to be in sync with shared/utils/const.ts
     base2: #98fb98,
     button: #32cd32,
     button-hover: #228b22
