+++ src/shared/assets/footer/github.png
