+++ src/shared/assets/footer/profile.jpg
