+++ src/versions/v2/layouts/Footer/_styles.css
@@ -1,0 +1,3 @@
+.ic img {
+  @apply h-3 w-3 mr-2 cursor-pointer object-contain last:mr-0;
+}
