+++ src/versions/v2/layouts/Layout/index.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Outlet } from 'react-router-dom'
+
+import Header from '../Header'
+import Footer from '../Footer'
+
+const Layout: React.FC = () => (
+  <div className="flex min-h-screen flex-col bg-white">
+    <Header />
+    <main className="flex flex-1 flex-col">
+      <Outlet />
+    </main>
+    <Footer />
+  </div>
+)
+
+export default Layout
