+++ src/versions/v2/utils/paths.ts
@@ -1,0 +1,7 @@
+/** Default site (v2). Hash: `#/`, `#/food-album`, `#/collections/:id` */
+export const V2_HOME = '/'
+
+export const v2FoodAlbumPath = '/food-album'
+
+export const v2CollectionPath = (collectionId: string) =>
+  `/collections/${collectionId}`
