+++ tailwind.config.js
@@ -2,12 +2,12 @@
 module.exports = {
   content: ["./src/**/*.{html,js,jsx,ts,tsx}"],
   theme: {
-    spacing: {
-      'header-height': '90px',
-      'header-padding': '5px',
-      'footer-height': '145px'
-    },
     extend: {
+      spacing: {
+        'header-height': '90px',
+        'header-padding': '5px',
+        'footer-height': '145px',
+      },
       minHeight: {
         header: '90px',
         content: 'calc(100vh - theme("spacing.header-height") - 1px - (theme("spacing.header-padding") * 2) - theme("spacing.footer-height") - 1px)'
